weight initialization project
In this lesson, I have learned how to find good initial weights for a neural network using PyTorch. Weight initialization happens once, when a model is created and before it trains. Having good initial weights can place the neural network close to the optimal solution. This allows the neural network to come to the best solution quicker.